Modified every file, added MAZES.py
Modified HELPER.py to contain 2 kinds fo classes, one for the maze
generator algorithms, and another for the visualization and pathfinding.
Re-did GUI.py, for now it just contains the main menu, and a basic main
loop.
Added MAZES.py, file that contains all the maze generator algorithms so far.